feat(backend): implement 2-layer cost+token quota enforcement
Implement cost-based quota system with token fallback:
- Add cost columns to agent_token_daily_usage model + migration
- Expand OpenClaw parser to extract totalCost and missingCostEntries
- Remove 0.5x billing multiplier, use raw tokens
- Implement 2-layer enforcement: cost primary, token safety net
- Board leads track usage but never block (observe-only)
- Add cost fields to EntitlementPolicy, AgentRead, QuotaUsage schemas
- Update plan tier limits for trial and pro

feat(frontend): display cost-based quota in UI
- Add cost fields to AgentRead type
- Update AgentsTable to show cost as primary, fallback to tokens
- Update QuotaSummary with USD formatting for cost resources

test: add cost-based enforcement tests
- Parser tests for cost field extraction
- 2-layer enforcement tests
- Update existing tests for raw token expectations

+++ backend/app/schemas/entitlements.py
@@ -38,9 +38,9 @@ class QuotaUsage(SQLModel):
     """Usage and capacity details for a single quota dimension."""
 
     resource: str
-    used: int
-    limit: int | None = None
-    remaining: int | None = None
+    used: float
+    limit: float | None = None
+    remaining: float | None = None
     exceeded: bool = False
